pyproject.toml
[tool.poetry]
name = "bottec-test"
version = "0.1.0"
description = ""
authors = ["anton <angeloffree@yandex.ru>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
coloredlogs = "^15.0.1"
python-dotenv = "^1.0.0"
alembic = "^1.10.4"
asyncpg = "^0.27.0"
sqlalchemy = "2.0"
psycopg2-binary = "^2.9.6"
greenlet = "^2.0.2"
httpx = "^0.24.0"
tenacity = "^8.2.2"
pytest-httpx = "^0.22.0"
pytest-asyncio = "^0.21.0"
aiosqlite = "^0.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
